+++ packages/extension/utils/checkShopl.ts
@@ -1,0 +1,6 @@
+export const checkShopl = (origin: string) => {
+  if (origin.includes('shoplworks') || origin.includes('localhost')) {
+    return true;
+  }
+  return false;
+};